steam: Remove environment trigger for the mangoapp implicit layer
This environment variable was used to enable the mangoapp
Vulkan implicit layer [1]. The mangoapp layer detects the
engine used by the application and reports it back to
mangoapp via a JSON file [2].

Currently reading the engine type is disabled [3] in mangoapp
and the environment variable was removed from the upstream
gamescope-session script. The implicit layer was never active
in our setup to begin with anyways (not added to the search
path).

[1] https://github.com/flightlessmango/MangoHud/blob/fc31ac626434e9fcbe4a071e16351722224f44a2/src/app/layer.json.in
[2] https://github.com/flightlessmango/MangoHud/blob/fc31ac626434e9fcbe4a071e16351722224f44a2/src/app/layer.cpp
[3] https://github.com/flightlessmango/MangoHud/commit/4eb6c83d81ea14a0430d48662b2c992e950afabd